requirements.txt
Django==4.2.13
django-model-utils==4.5.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
pytest==8.2.1
